update-16.12

+++ app/Http/Controllers/Admin/CategoriesController.php
@@ -32,7 +32,24 @@ class CategoriesController extends Controller
      */
     public function store(Request $request)
     {
-        Category::create($request->all()); 
+        $data = $request->all();
+
+        if(empty($request->image_default)){
+
+            $data['image'] = null;
+
+        }
+        else if($request->file('image')){
+       
+            $image = $request->file('image');
+            $customer_image = $image->getClientOriginalName();
+            $destinationPath = public_path('/backend/images/categories/');
+            $image->move($destinationPath,$customer_image);
+
+            $data['image'] = '/backend/images/categories/'.$customer_image;
+        }
+
+        Category::create($data); 
         return redirect()->route('category.index')->with('success', 'Thêm mới thành công');;
     }
 
@@ -61,12 +78,26 @@ class CategoriesController extends Controller
       
         $data = $request->all();
         $data['is_active'] = $request->is_active ? 1 : 0;
+
         $cate = Category::find($id);
-        if(empty($request->image)){
-            if(!empty($cate->image)){
-                File::delete(public_path($cate->image));
-            }
+
+        if(!empty($cate->image)){
+            File::delete(public_path($cate->image));
         }
+
+        if(empty($request->image_default)){
+            $data['image'] = null;
+        }
+        else if($request->file('image')){
+       
+            $image = $request->file('image');
+            $customer_image = $image->getClientOriginalName();
+            $destinationPath = public_path('/backend/images/categories/');
+            $image->move($destinationPath,$customer_image);
+
+            $data['image'] = '/backend/images/categories/'.$customer_image;
+        }
+
         $cate->update($data);
         return redirect()->route('category.edit',$id)->with('success', 'Cập nhật thành công');
     }
@@ -77,6 +108,9 @@ class CategoriesController extends Controller
     public function destroy($id)
     {
         $cate = Category::find($id);
+        if(!empty($cate->product->id)){
+            return redirect()->route('category.index')->with('error', 'Danh mục đang tồn tại sản phẩm');
+        }
         $cate->delete(); 
         return redirect()->route('category.index')->with('success', 'Xóa thành công');
     }
